#12 - optimize handle errors
