+++ src/constants/httpStatus.ts
@@ -11,4 +11,7 @@ const HTTP_STATUS = {
   INTERNAL_SERVER_ERROR: 500
 } as const
 
+export type HttpStatusCode = (typeof HTTP_STATUS)[keyof typeof HTTP_STATUS]
+type HttpStatusKey = keyof typeof HTTP_STATUS
+
 export default HTTP_STATUS
