+++ src/middleware/errorHandler.middleware.ts
@@ -1,4 +1,6 @@
+import HTTP_STATUS from '@/constants/httpStatus'
 import { Request, Response, NextFunction } from 'express'
+import { omit } from 'lodash'
 
 export const badRequestException = (req: Request, res: Response, next: NextFunction) => {
   const error = new Error(`Route ${req.originalUrl} not found.`)
@@ -7,7 +9,12 @@ export const badRequestException = (req: Request, res: Response, next: NextFunct
 }
 
 export const defaultErrorHandler = (err: any, req: Request, res: Response, next: NextFunction) => {
-  return res.status(400).json({
-    err
+  Object.getOwnPropertyNames(err).forEach((key) => {
+    Object.defineProperty(err, key, { enumerable: true })
+  })
+
+  res.status(HTTP_STATUS.INTERNAL_SERVER_ERROR).json({
+    message: err.message,
+    errorInfo: omit(err, ['stack'])
   })
 }
